requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
yt-dlp==2023.10.13
Pillow==10.0.1
python-decouple==3.8
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0